requirements.txt
alembic==1.0.3
astroid==2.0.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-filter==2.0.0
django-heroku==0.3.1
django-pyodbc-azure==2.1.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
Flask==1.0.2
Flask-Migrate==2.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2==2.7.6.1
pylint==2.1.1
pylint-django==2.0.4
pylint-plugin-utils==0.4
pyodbc==4.0.24
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
requests==2.20.1
simplejson==3.16.0
six==1.11.0
SQLAlchemy==1.2.14
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.1.0
Werkzeug==0.14.1
whitenoise==4.1.2
wrapt==1.10.11
